+++ src/main/java/sesac/server/user/controller/UserController.java
@@ -1,6 +1,5 @@
 package sesac.server.user.controller;
 
-import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -8,7 +7,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.query.Param;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -22,17 +20,13 @@ import org.springframework.web.bind.annotation.RestController;
 import sesac.server.auth.dto.AuthPrincipal;
 import sesac.server.auth.dto.CustomPrincipal;
 import sesac.server.common.dto.PageResponse;
-import sesac.server.common.exception.BindingResultHandler;
 import sesac.server.user.dto.request.AcceptStatusRequest;
-import sesac.server.user.dto.request.MessageSendRequest;
 import sesac.server.user.dto.request.SearchStudentRequest;
 import sesac.server.user.dto.request.UpdateStudentRequest;
 import sesac.server.user.dto.response.ManagerListResponse;
-import sesac.server.user.dto.response.MessageResponse;
 import sesac.server.user.dto.response.SearchStudentResponse;
 import sesac.server.user.dto.response.StudentDetailResponse;
 import sesac.server.user.dto.response.StudentListResponse;
-import sesac.server.user.exception.UserErrorCode;
 import sesac.server.user.service.UserService;
 
 @Log4j2
@@ -66,59 +60,6 @@ public class UserController {
         return null;
     }
 
-    // -----------------------------------------------------------쪽지
-    @GetMapping("messages/received")
-    public ResponseEntity<List<MessageResponse>> getReceivedMessageList(
-            @AuthPrincipal CustomPrincipal user,
-            @PageableDefault Pageable pageable
-    ) {
-        List<MessageResponse> response = userService.receivedMessage(user.id(), pageable);
-        return ResponseEntity.ok().body(response);
-    }
-
-
-    @GetMapping("messages/sent")
-    public ResponseEntity<List<MessageResponse>> getSentMessageList(
-            @AuthPrincipal CustomPrincipal user,
-            @PageableDefault Pageable pageable
-    ) {
-        List<MessageResponse> response = userService.sentMessage(user.id(), pageable);
-        return ResponseEntity.ok().body(response);
-    }
-
-    @PostMapping("messages/{userId}")
-    public ResponseEntity<Void> sendMessage(
-            @AuthPrincipal CustomPrincipal sender,
-            @PathVariable Long userId,
-            @Valid @RequestBody MessageSendRequest request,
-            BindingResult bindingResult
-    ) {
-        BindingResultHandler.handle(bindingResult, List.of(UserErrorCode.REQUIRED_MESSAGE));
-
-        userService.sendMessage(sender.id(), userId, request);
-
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("messages/{messageId}")
-    public ResponseEntity<MessageResponse> getMessage(
-            @AuthPrincipal CustomPrincipal user,
-            @PathVariable Long messageId
-    ) {
-        MessageResponse response = userService.getMessage(user.id(), messageId);
-
-        return ResponseEntity.ok().body(response);
-    }
-
-    @DeleteMapping("messages/{messageId}")
-    public ResponseEntity<Void> deleteMessage(
-            @AuthPrincipal CustomPrincipal receiver,
-            @PathVariable Long messageId
-    ) {
-        userService.deleteMessage(receiver.id(), messageId);
-
-        return ResponseEntity.ok().build();
-    }
 
     // -----------------------------------------------------------프로필
     @PutMapping("profiles")
